+++ src/main/java/ru/Nik_Russkikh/BooksApp/repository/BookRepository.java
@@ -5,7 +5,7 @@ import org.springframework.data.r2dbc.repository.R2dbcRepository;
 import reactor.core.publisher.Flux;
 import ru.Nik_Russkikh.BooksApp.model.Book;
 
-public interface BookRepository extends R2dbcRepository<Book, Integer> {
+public interface BookRepository extends R2dbcRepository<Book, Long> {
 
     Flux<Book> findByAuthor(String author);
